+++ explorer/client/src/pages/object-result/ObjectResult.tsx
@@ -4,7 +4,7 @@ import { useParams } from 'react-router-dom';
 import ErrorResult from '../../components/error-result/ErrorResult';
 import Longtext from '../../components/longtext/Longtext';
 import theme from '../../styles/theme.module.css';
-import { type AddressOwner, DefaultRpcClient } from '../../utils/rpc';
+import { type AddressOwner, DefaultRpcClient } from '../../utils/SuiRpcClient';
 import {asciiFromNumberBytes, trimStdLibPrefix } from '../../utils/stringUtils';
 
 import styles from './ObjectResult.module.css';
